feat[Yahoo_DSP]: refactor code
